requirements.txt
discord.py==2.3.2
requests==2.32.4
python-dotenv==1.0.1
pytz==2024.1
chat-exporter==2.7.1
pillow==8.4.0
python-dateutil==2.9.0
google-auth==2.37.0
google-api-python-client==2.172.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
python-dotenv